vasarely_service/requirements.txt
flask
pyfiglet
Pillow